+++ notify/slack/types.go
@@ -19,14 +19,13 @@ import (
 	"log/slog"
 	"net/http"
 
-	"github.com/prometheus/alertmanager/config"
 	"github.com/prometheus/alertmanager/notify"
 	"github.com/prometheus/alertmanager/template"
 )
 
 // Notifier implements a Notifier for Slack notifications.
 type Notifier struct {
-	conf    *config.SlackConfig
+	conf    *Config
 	tmpl    *template.Template
 	logger  *slog.Logger
 	client  *http.Client
@@ -49,19 +48,19 @@ type request struct {
 
 // attachment is used to display a richly formatted message block.
 type attachment struct {
-	Title      string               `json:"title,omitempty"`
-	TitleLink  string               `json:"title_link,omitempty"`
-	Pretext    string               `json:"pretext,omitempty"`
-	Text       string               `json:"text"`
-	Fallback   string               `json:"fallback"`
-	CallbackID string               `json:"callback_id"`
-	Fields     []config.SlackField  `json:"fields,omitempty"`
-	Actions    []config.SlackAction `json:"actions,omitempty"`
-	ImageURL   string               `json:"image_url,omitempty"`
-	ThumbURL   string               `json:"thumb_url,omitempty"`
-	Footer     string               `json:"footer"`
-	Color      string               `json:"color,omitempty"`
-	MrkdwnIn   []string             `json:"mrkdwn_in,omitempty"`
+	Title      string   `json:"title,omitempty"`
+	TitleLink  string   `json:"title_link,omitempty"`
+	Pretext    string   `json:"pretext,omitempty"`
+	Text       string   `json:"text"`
+	Fallback   string   `json:"fallback"`
+	CallbackID string   `json:"callback_id"`
+	Fields     []Field  `json:"fields,omitempty"`
+	Actions    []Action `json:"actions,omitempty"`
+	ImageURL   string   `json:"image_url,omitempty"`
+	ThumbURL   string   `json:"thumb_url,omitempty"`
+	Footer     string   `json:"footer"`
+	Color      string   `json:"color,omitempty"`
+	MrkdwnIn   []string `json:"mrkdwn_in,omitempty"`
 }
 
 // slackResponse represents the response from Slack API.
